updated ui for admin products page

+++ src/components/admin/dashboard/views/AdminDashboard.vue
@@ -1,110 +1,33 @@
 <template>
   <NavBar />
-  <aside
-    id="logo-sidebar"
-    class="fixed top-0 left-0 z-40 w-64 h-screen pt-20 transition-transform -translate-x-full bg-secondary border-accent border-r sm:translate-x-0"
-    aria-label="Sidebar"
-  >
-    <div class="h-full px-3 pb-4 overflow-y-auto bg-secondary border-accent">
-      <ul class="space-y-2 font-medium">
-        <li>
-          <router-link
-            to="/admin"
-            class="flex items-center p-2 text-secondary-foreground rounded-lg hover:bg-primary/10 group"
-          >
-            <svg
-              class="w-5 h-5 text-secondary-foreground transition duration-75 group-hover:text-primary"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              viewBox="0 0 22 21"
-            >
-              <path
-                d="M16.975 11H10V4.025a1 1 0 0 0-1.066-.998 8.5 8.5 0 1 0 9.039 9.039.999.999 0 0 0-1-1.066h.002Z"
-              />
-              <path
-                d="M12.5 0c-.157 0-.311.01-.565.027A1 1 0 0 0 11 1.02V10h8.975a1 1 0 0 0 1-.935c.013-.188.028-.374.028-.565A8.51 8.51 0 0 0 12.5 0Z"
-              />
-            </svg>
-            <span class="ms-3">Dashboard</span>
-          </router-link>
-        </li>
-
-        <li>
-          <a
-            href="#"
-            class="flex items-center p-2 text-secondary-foreground rounded-lg hover:bg-primary/10 group"
-          >
-            <svg
-              class="flex-shrink-0 w-5 h-5 text-secondary-foreground transition duration-75 group-hover:text-primary"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              viewBox="0 0 20 20"
-            >
-              <path
-                d="m17.418 3.623-.018-.008a6.713 6.713 0 0 0-2.4-.569V2h1a1 1 0 1 0 0-2h-2a1 1 0 0 0-1 1v2H9.89A6.977 6.977 0 0 1 12 8v5h-2V8A5 5 0 1 0 0 8v6a1 1 0 0 0 1 1h8v4a1 1 0 0 0 1 1h2a1 1 0 0 0 1-1v-4h6a1 1 0 0 0 1-1V8a5 5 0 0 0-2.582-4.377ZM6 12H4a1 1 0 0 1 0-2h2a1 1 0 0 1 0 2Z"
-              />
-            </svg>
-            <span class="flex-1 ms-3 whitespace-nowrap">Inbox</span>
-            <span
-              class="inline-flex items-center justify-center w-3 h-3 p-3 ms-3 text-sm font-medium text-primary-foreground bg-primary rounded-full"
-              >3</span
-            >
-          </a>
-        </li>
-        <li>
-          <a
-            href="#"
-            class="flex items-center p-2 text-secondary-foreground rounded-lg hover:bg-primary/10 group"
-          >
-            <svg
-              class="flex-shrink-0 w-5 h-5 text-secondary-foreground transition duration-75 group-hover:text-primary"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              viewBox="0 0 20 18"
-            >
-              <path
-                d="M14 2a3.963 3.963 0 0 0-1.4.267 6.439 6.439 0 0 1-1.331 6.638A4 4 0 1 0 14 2Zm1 9h-1.264A6.957 6.957 0 0 1 15 15v2a2.97 2.97 0 0 1-.184 1H19a1 1 0 0 0 1-1v-1a5.006 5.006 0 0 0-5-5ZM6.5 9a4.5 4.5 0 1 0 0-9 4.5 4.5 0 0 0 0 9ZM8 10H5a5.006 5.006 0 0 0-5 5v2a1 1 0 0 0 1 1h11a1 1 0 0 0 1-1v-2a5.006 5.006 0 0 0-5-5Z"
-              />
-            </svg>
-            <span class="flex-1 ms-3 whitespace-nowrap">Users</span>
-          </a>
-        </li>
-        <li>
-          <a
-            href="#"
-            class="flex items-center p-2 text-secondary-foreground rounded-lg hover:bg-primary/10 group"
-          >
-            <svg
-              class="flex-shrink-0 w-5 h-5 text-secondary-foreground transition duration-75 group-hover:text-primary"
-              aria-hidden="true"
-              xmlns="http://www.w3.org/2000/svg"
-              fill="currentColor"
-              viewBox="0 0 18 20"
-            >
-              <path
-                d="M17 5.923A1 1 0 0 0 16 5h-3V4a4 4 0 1 0-8 0v1H2a1 1 0 0 0-1 .923L.086 17.846A2 2 0 0 0 2.08 20h13.84a2 2 0 0 0 1.994-2.153L17 5.923ZM7 9a1 1 0 0 1-2 0V7h2v2Zm0-5a2 2 0 1 1 4 0v1H7V4Zm6 5a1 1 0 1 1-2 0V7h2v2Z"
-              />
-            </svg>
-            <span class="flex-1 ms-3 whitespace-nowrap">Products</span>
-          </a>
-        </li>
-      </ul>
-    </div>
-  </aside>
+  <AdminSidebar />
 
   <div class="p-4 ml-2 sm:ml-64">
-    <div class="flex flex-col p-4 border-2 rounded-lg py-14">
-      <div class="flex flex-row">
+    <div class="flex flex-row justify-start py-10">
+      <div class="flex">
+        <span class="material-symbols-outlined py-2 px-2 text-5xl">
+          dashboard
+        </span>
+      </div>
+      <div class="flex flex-col">
+        <h1 class="text-2xl">Welcome to the Admin Dashboard</h1>
+        <p class="text-sm py-1">
+          This is the admin dashboard. You can manage your products, users, and
+          other settings from here.
+        </p>
+      </div>
+    </div>
+    <div class="flex flex-col p-4 border-2 rounded-lg py-5">
+      <div class="flex flex-row border-b-2">
         <h1 class="font-bold text-xl tracking-wide">Featured Images</h1>
         <div class="p-1">
           <!-- Upload Featured Image -->
           <Popover v-model="isUploadPopoverOpen">
             <PopoverTrigger>
-              <button title="Add a Featured Image">
-                <span class="material-symbols-outlined"> add_box </span>
+              <button title="Add an Image to be Featured">
+                <span class="material-symbols-outlined text-emerald-300">
+                  add_box
+                </span>
               </button>
             </PopoverTrigger>
             <PopoverContent>
@@ -125,7 +48,6 @@
                     <AlertDialogTrigger>
                       <button
                         type="submit"
-                        @click=""
                         class="border-2 bg-secondary border-horder rounded-lg p-2 text-sm text-secondary-foreground"
                       >
                         Add Featured Image
@@ -161,13 +83,37 @@
           class="relative flex items-end"
         >
           <img class="h-auto max-w-full rounded-lg" :src="url" alt="" />
-          <button
-            class="absolute text-end p-3 text-destructive hover:text-destructive-foreground hover:bg-destructive hover:opacity-90 w-full"
-            title="Delete Image"
-            @click="() => deleteFeaturedImage(url)"
-          >
-            <span class="material-symbols-outlined"> delete </span>
-          </button>
+          <AlertDialog>
+            <AlertDialogTrigger class="absolute w-full text-end"
+              ><button
+                class="p-3 hover:text-destructive text-destructive-foreground bg-destructive rounded-br-lg hover:bg-gray-200"
+                title="Delete Image"
+              >
+                <span class="material-symbols-outlined"> delete </span>
+              </button></AlertDialogTrigger
+            >
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Delete Image</AlertDialogTitle>
+                <AlertDialogDescription>
+                  Are you sure you want to delete this image?
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+              <AlertDialogFooter>
+                <AlertDialogAction
+                  @click="() => deleteFeaturedImage(url)"
+                  class="bg-destructive text-destructive-foreground"
+                >
+                  Delete
+                </AlertDialogAction>
+                <AlertDialogAction
+                  class="bg-secondary text-secondary-foreground"
+                >
+                  Cancel
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       </div>
     </div>
@@ -175,7 +121,7 @@
   <div class="p-4 ml-2 sm:ml-64">
     <div class="flex flex-col p-4 border-2 rounded-lg py-14">
       <div class="flex flex-row">
-        <h1 class="font-bold text-xl tracking-wide">Products Images</h1>
+        <h1 class="font-bold text-xl tracking-wide">Products Statistics</h1>
       </div>
       <div class="grid grid-cols-2 md:grid-cols-3 gap-4 p-5">
         <div class="">
@@ -212,6 +158,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { setup as setupControllerFeatured } from "../controllers/adminUploadFeatured.ts";
 import { useRouter } from "vue-router";
+import AdminSidebar from "../views/AdminSidebar.vue";
 
 const { onFileChange, uploadFeaturedImage, imageUrls, deleteFeaturedImage } =
   setupControllerFeatured();
